+++ ft_putnbr.c
@@ -6,7 +6,7 @@
 /*   By: amagnell <amagnell@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/18 14:07:08 by amagnell          #+#    #+#             */
-/*   Updated: 2023/06/28 19:10:09 by amagnell         ###   ########.fr       */
+/*   Updated: 2023/10/14 20:18:38 by amagnell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,14 +21,14 @@ int	ft_putnbr(int count, long long nbr, char *base, const char *type)
 		if (count == -1)
 			return (-1);
 	}
-	if (nbr >= ft_strlen(base))
+	if (nbr >= (int)ft_strlen(base))
 	{
 		count = ft_putnbr(count, nbr / ft_strlen(base), base, type);
 		if (count == -1)
 			return (-1);
 		count = ft_putnbr(count, nbr % ft_strlen(base), base, type);
 	}
-	else if (nbr < ft_strlen(base))
+	else if (nbr < (int)ft_strlen(base))
 		count = count + ft_putchar(base[nbr]);
 	return (count);
 }
